refactor(prs): apply the quality findings, and make /quality require it
The four findings from the PR #988 review were verified as real and then
left in the code. Applying them:

- The two `deriveGithubSnapshot*` mappers capture nothing from the
  prService factory — they are pure row->DTO functions and now live at
  module scope beside `rowToSummary`, where they can be tested without
  constructing a service.
- The sticky period header is derived during render from the range being
  drawn. It previously round-tripped through a ref and a state-setting
  effect, re-rendering the whole list every scroll frame to compute
  something already in hand.
- `detachRows` takes an id list instead of a SQL predicate string plus a
  params array that four separate statements had to agree with
  positionally.
- The GitHub PR row renderer moves to its own module. None of it depends
  on GitHubTab's state, so the tab is left as a coordinator: 2468 -> 1960
  lines.

The skill changes matter more than the refactor. /quality's synthesis
step said "apply the safe fixes ... gate the rest", which reads as
licence to verify a finding and then defer it — so a run could surface
real problems, fix none of them, and still report success. It now says
fix every verified finding at any severity, and narrows the gate to the
only two things that genuinely need the author: a product decision, or a
behavior change the branch was not asked to make. "Structural", "large"
and "out of scope" are named as non-reasons.

/ship gained the matching precondition: a non-empty quality gate blocks
the merge, rather than being mentioned afterwards.
