+++ apps/desktop/src/renderer/components/prs/tabs/githubPrBranch.ts
@@ -1,0 +1,4 @@
+/** Strip the `refs/heads/` prefix so lane refs and PR head branches compare directly. */
+export function branchNameFromRef(ref: string | null | undefined): string {
+  return String(ref ?? "").replace(/^refs\/heads\//, "").trim();
+}
